requirements.txt
requests>=2.32.3
allure-pytest>=2.14.2
pytest>=8.4.0
selenium>=4.33.0
pytest-html>=4.1.1
webdriver-manager>=4.0.2